move music player to seed 24 components
