+++ data/seed/24/index.ts
@@ -1,4 +1,4 @@
-import { TrackInfo } from '@components/music-player';
+import { TrackInfo } from '@components/seed/24/music-player';
 import data from './data.json';
 
 export const seed24AudioData: TrackInfo[] = data;
