+++ src/components/seed/24/music-player/NowPlaying.tsx
@@ -1,9 +1,9 @@
 import { Stack } from '@mui/material';
 import { Box } from '@mui/system';
-import PlayerTrackInformation from '@components/music-player/PlayerTrackInformation';
-import PlayerTimeSlider from '@components/music-player/PlayerTimeSlider';
-import PlayerButtons from '@components/music-player/PlayerButtons';
-import PlayerVolumeSlider from '@components/music-player/PlayerVolumeSlider';
+import PlayerTrackInformation from '@components/seed/24/music-player/PlayerTrackInformation';
+import PlayerTimeSlider from '@components/seed/24/music-player/PlayerTimeSlider';
+import PlayerButtons from '@components/seed/24/music-player/PlayerButtons';
+import PlayerVolumeSlider from '@components/seed/24/music-player/PlayerVolumeSlider';
 import { isMobile } from '@tools/helper';
